% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mawasu.R
\name{rendermawasu}
\alias{rendermawasu}
\title{Widget render function for use in Shiny}
\usage{
rendermawasu(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

